Move fixtures to separate file
